rip bobcatutil :(
