+++ src/main/java/frc/lib/BobcatLib/Util/DSUtil.java
@@ -1,0 +1,19 @@
+package frc.lib.BobcatLib.Util;
+
+import edu.wpi.first.wpilibj.DriverStation;
+import edu.wpi.first.wpilibj.DriverStation.Alliance;
+
+public class DSUtil {
+    public static Alliance getAlliance() {
+        return DriverStation.getAlliance().isEmpty() ? Alliance.Blue : DriverStation.getAlliance().get();
+    }
+
+    public static boolean isBlue() {
+        return getAlliance() == Alliance.Blue;
+    }
+
+    public static boolean isRed() {
+        return getAlliance() == Alliance.Red;
+    }
+    
+}
